+++ app/proguard-rules.pro
@@ -19,11 +19,17 @@
 # If you keep the line number information, uncomment this to
 # hide the original source file name.
 #-renamesourcefileattribute SourceFile
--keep class org.bouncycastle.** { *; }
--dontwarn org.bouncycastle.**
 
--keep class org.conscrypt.** { *; }
--dontwarn org.conscrypt.**
+-keep class kr.nbc.momo.** { *; }
+#-dontwarn kr.nbc.momo.**
 
--keep class org.openjsse.** { *; }
--dontwarn org.openjsse.**
+# missing rules
+-dontwarn org.bouncycastle.jsse.BCSSLParameters
+-dontwarn org.bouncycastle.jsse.BCSSLSocket
+-dontwarn org.bouncycastle.jsse.provider.BouncyCastleJsseProvider
+-dontwarn org.conscrypt.Conscrypt$Version
+-dontwarn org.conscrypt.Conscrypt
+-dontwarn org.conscrypt.ConscryptHostnameVerifier
+-dontwarn org.openjsse.javax.net.ssl.SSLParameters
+-dontwarn org.openjsse.javax.net.ssl.SSLSocket
+-dontwarn org.openjsse.net.ssl.OpenJSSE
