--- app/src/main/java/kr/nbc/momo/presentation/mypage/ImageEnumClass.kt
@@ -1,5 +1,0 @@
-package kr.nbc.momo.presentation.mypage
-
-enum class ImageType {
-    PROFILE, BACKGROUND, PORTFOLIO
-}
