+++ app/src/main/java/kr/nbc/momo/presentation/mypage/group/adapter/MemberGroupAdapter.kt
@@ -1,4 +1,4 @@
-package kr.nbc.momo.presentation.mypage.adapter
+package kr.nbc.momo.presentation.mypage.group.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,7 +6,8 @@ import androidx.recyclerview.widget.RecyclerView
 import kr.nbc.momo.databinding.RvItemHomeHorizontalBinding
 import kr.nbc.momo.presentation.group.model.GroupModel
 import kr.nbc.momo.util.setThumbnailByUrlOrDefault
-class LeaderGroupAdapter(private var items: List<GroupModel>): RecyclerView.Adapter<LeaderGroupAdapter.Holder>() {
+
+class MemberGroupAdapter(private var items: List<GroupModel>): RecyclerView.Adapter<MemberGroupAdapter.Holder>() {
     interface ItemClick{
         fun itemClick(position: Int)
     }
