+++ app/src/main/res/layout/fragment_my_group.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/grey"
-    tools:context=".presentation.mypage.MyGroupFragment">
+    tools:context=".presentation.mypage.group.MyGroupFragment">
 
     <TextView
         android:id="@+id/tvGroupLeader"
