+++ app/src/main/res/layout/fragment_my_page_container.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.mypage.MyPageContainerFragment">
+    tools:context=".presentation.mypage.profile.MyPageContainerFragment">
 
     <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:id="@+id/fragment_container"
